# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Superhéroes - Sistema de Batallas por Equipos",
    "version": "4.0.0",
    "description": "API para gestión de superhéroes, villanos y batallas por equipos con persistencia en MongoDB Atlas. **Nueva versión con autenticación JWT**: cada usuario tiene sus propias batallas privadas. Los endpoints de batalla requieren autenticación con token Bearer. Incluye sistema de turnos, IDs únicos, barras de poder/defensa, y validaciones mejoradas para batallas estratégicas."
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Servidor local"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Registrar nuevo usuario",
        "description": "Crea una nueva cuenta de usuario con email y contraseña",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string", "example": "usuario1" },
                  "email": { "type": "string", "format": "email", "example": "usuario@email.com" },
                  "password": { "type": "string", "minLength": 6, "example": "password123" }
                },
                "required": ["username", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "username": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Error de validación o usuario ya existe" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Iniciar sesión",
        "description": "Autentica un usuario y devuelve un token JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "usuario@email.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "username": { "type": "string" },
                        "email": { "type": "string" }
                      }
                    },
                    "token": { "type": "string" }
                  }
                }
              }
            }
          },
          "401": { "description": "Credenciales inválidas" }
        }
      }
    },
    "/api/auth/profile": {
      "get": {
        "tags": ["Autenticación"],
        "summary": "Obtener perfil del usuario autenticado",
        "description": "Devuelve la información del usuario autenticado",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Perfil del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "username": { "type": "string" },
                        "email": { "type": "string" },
                        "createdAt": { "type": "string", "format": "date-time" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Token no válido o expirado" }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Autenticación"],
        "summary": "Cerrar sesión",
        "description": "Cierra la sesión del usuario (solo confirmación, el token expira naturalmente)",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "Logout exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Pega solo el token JWT (sin 'Bearer'). El sistema lo agregará automáticamente."
      }
    },
...
  }
}